+++ tests/phpunit/tests/linkify-categories.php
@@ -155,16 +155,4 @@ class Linkify_Categories_Test extends WP_UnitTestCase {
 		$this->assertEquals( $expected, $this->get_results( array( array(), '<ul><li>', '</li></ul>', '</li><li>', '', $missing ), false ) );
 	}
 
-	/**
-	 * @expectedDeprecated linkify_categories
-	 */
-	public function test_deprecated_function() {
-		$this->assertEquals( $this->expected_output( 1 ), $this->get_results( array( $this->cat_ids[0] ), false, true ) );
-		$this->assertEquals( $this->expected_output( 5 ), $this->get_results( array( $this->cat_ids ), false, true ) );
-		$cat = get_category( $this->cat_ids[0] );
-		$this->assertEquals( $this->expected_output( 1 ), $this->get_results( array( $cat->slug ), false, true ) );
-		$cat_slugs = array_map( array( $this, 'get_slug' ), $this->cat_ids );
-		$this->assertEquals( $this->expected_output( 5 ), $this->get_results( array( $cat_slugs ), false, true ) );
-	}
-
 }